requirements.txt
python-mpd2
gpiozero
rpi-ws281x
pyserial
lgpio